# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DigiPulse API",
        "description": "API documentation for the DigiPulse monitoring system. **Authentication:** Use these endpoints to register, login, and manage user tokens.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/telegram/connect": {
            "get": {
                "tags": [
                    "Telegram"
                ],
                "summary": "Get Telegram connection link",
                "description": "Generate a connection token for the authenticated user.",
                "operationId": "e99c447db4a0c21e4a11160e27b635f8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "connected": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "random_32_char_token",
                                            "nullable": true
                                        },
                                        "bot_username": {
                                            "type": "string",
                                            "example": "DigiPulseBot",
                                            "nullable": true
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://t.me/DigiPulseBot?start=random_32_char_token",
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Telegram is already connected.",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "frontendKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/telegram/disconnect": {
            "post": {
                "tags": [
                    "Telegram"
                ],
                "summary": "Disconnect Telegram bot",
                "description": "Disconnect the Telegram bot for the authenticated user.",
                "operationId": "86f9c075e07f02b335f5368adacba446",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "frontendKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserSchema": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "first_name",
                    "last_name",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "johndoug"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johtdouj@gmail.pro"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-16T09:01:11.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-16T09:01:11.000000Z"
                    },
                    "telegram_chat_id": {
                        "type": "string",
                        "example": "123456789",
                        "nullable": true
                    },
                    "is_telegram_connected": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "frontendKey": {
                "type": "apiKey",
                "description": "Custom frontend key for API access",
                "name": "X-Frontend-Key",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "BearerAuth access token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}